+++ example/example/urls.py
@@ -1,6 +1,8 @@
 from django.conf.urls import include, url
+from django.contrib import admin
 
 
 urlpatterns = [
+    url('^admin/', include(admin.site.urls)),
     url(r'', include('conman.routes.urls')),
 ]